Started chaign RobotContainer
